Fix Cherry-Pick conflicts
